+++ Changelog.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+-   Property `apps` in type `SourceJson` became mandatory.
+
 ## 6.13.2 - 2022-12-19
 
 ## Changed