+++ Changelog.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 6.14.0 - 2023-01-02
+
+### Changed
+
+-   Property `apps` in type `SourceJson` became mandatory.
+-   Renamed type `AppJson` to `AppInfo`.
+-   Renamed property `latest` to `latestVersion` in type `AppInfo`.
+-   Exported type `AppVersions` and renamed property `tarball` to `tarballUrl`.
+-   Bumped target for TypeScript to `ES2020`.
+
 ## 6.13.2 - 2022-12-19
 
 ## Changed